# Fix ActiveRecord cache on queries with limit/offset