--- a/activerecord/test/cases/collection_cache_key_test.rb
+++ b/activerecord/test/cases/collection_cache_key_test.rb
@@ -13,60 +13,68 @@ module ActiveRecord
     fixtures :developers, :projects, :developers_projects, :topics, :comments, :posts
 
     test "collection_cache_key on model" do
-      assert_match(/\Adevelopers\/query-(\h+)-(\d+)-(\d+)\z/, Developer.collection_cache_key)
+      assert_match(/\Adevelopers\/query-(\h+)-(\d+)-(\d+)-(\d+)\z/, Developer.collection_cache_key)
     end
 
     test "cache_key for relation" do
       developers = Developer.where(salary: 100000).order(updated_at: :desc)
       last_developer_timestamp = developers.first.updated_at
+      first_developer_timestamp = developers.last.updated_at
 
-      assert_match(/\Adevelopers\/query-(\h+)-(\d+)-(\d+)\z/, developers.cache_key)
+      assert_match(/\Adevelopers\/query-(\h+)-(\d+)-(\d+)-(\d+)\z/, developers.cache_key)
 
-      /\Adevelopers\/query-(\h+)-(\d+)-(\d+)\z/ =~ developers.cache_key
+      /\Adevelopers\/query-(\h+)-(\d+)-(\d+)-(\d+)\z/ =~ developers.cache_key
 
       assert_equal ActiveSupport::Digest.hexdigest(developers.to_sql), $1
       assert_equal developers.count.to_s, $2
       assert_equal last_developer_timestamp.to_s(ActiveRecord::Base.cache_timestamp_format), $3
+      assert_equal first_developer_timestamp.to_s(ActiveRecord::Base.cache_timestamp_format), $4
     end
 
     test "cache_key for relation with limit" do
       developers = Developer.where(salary: 100000).order(updated_at: :desc).limit(5)
       last_developer_timestamp = developers.first.updated_at
+      first_developer_timestamp = developers.last.updated_at
 
-      assert_match(/\Adevelopers\/query-(\h+)-(\d+)-(\d+)\z/, developers.cache_key)
+      assert_match(/\Adevelopers\/query-(\h+)-(\d+)-(\d+)-(\d+)\z/, developers.cache_key)
 
-      /\Adevelopers\/query-(\h+)-(\d+)-(\d+)\z/ =~ developers.cache_key
+      /\Adevelopers\/query-(\h+)-(\d+)-(\d+)-(\d+)\z/ =~ developers.cache_key
 
       assert_equal ActiveSupport::Digest.hexdigest(developers.to_sql), $1
       assert_equal developers.count.to_s, $2
       assert_equal last_developer_timestamp.to_s(ActiveRecord::Base.cache_timestamp_format), $3
+      assert_equal first_developer_timestamp.to_s(ActiveRecord::Base.cache_timestamp_format), $4
     end
 
     test "cache_key for relation with custom select and limit" do
       developers = Developer.where(salary: 100000).order(updated_at: :desc).limit(5)
       developers_with_select = developers.select("developers.*")
       last_developer_timestamp = developers.first.updated_at
+      first_developer_timestamp = developers.last.updated_at
 
-      assert_match(/\Adevelopers\/query-(\h+)-(\d+)-(\d+)\z/, developers_with_select.cache_key)
+      assert_match(/\Adevelopers\/query-(\h+)-(\d+)-(\d+)-(\d+)\z/, developers_with_select.cache_key)
 
-      /\Adevelopers\/query-(\h+)-(\d+)-(\d+)\z/ =~ developers_with_select.cache_key
+      /\Adevelopers\/query-(\h+)-(\d+)-(\d+)-(\d+)\z/ =~ developers_with_select.cache_key
 
       assert_equal ActiveSupport::Digest.hexdigest(developers_with_select.to_sql), $1
       assert_equal developers.count.to_s, $2
       assert_equal last_developer_timestamp.to_s(ActiveRecord::Base.cache_timestamp_format), $3
+      assert_equal first_developer_timestamp.to_s(ActiveRecord::Base.cache_timestamp_format), $4
     end
 
     test "cache_key for loaded relation" do
       developers = Developer.where(salary: 100000).order(updated_at: :desc).limit(5).load
       last_developer_timestamp = developers.first.updated_at
+      first_developer_timestamp = developers.last.updated_at
 
-      assert_match(/\Adevelopers\/query-(\h+)-(\d+)-(\d+)\z/, developers.cache_key)
+      assert_match(/\Adevelopers\/query-(\h+)-(\d+)-(\d+)-(\d+)\z/, developers.cache_key)
 
-      /\Adevelopers\/query-(\h+)-(\d+)-(\d+)\z/ =~ developers.cache_key
+      /\Adevelopers\/query-(\h+)-(\d+)-(\d+)-(\d+)\z/ =~ developers.cache_key
 
       assert_equal ActiveSupport::Digest.hexdigest(developers.to_sql), $1
       assert_equal developers.count.to_s, $2
       assert_equal last_developer_timestamp.to_s(ActiveRecord::Base.cache_timestamp_format), $3
+      assert_equal first_developer_timestamp.to_s(ActiveRecord::Base.cache_timestamp_format), $4
     end
 
     test "cache_key for relation with table alias" do
@@ -81,24 +89,26 @@ module ActiveRecord
       )
       developers = developers.where(salary: 100000).order(updated_at: :desc)
       last_developer_timestamp = developers.first.updated_at
+      first_developer_timestamp = developers.last.updated_at
 
-      assert_match(/\Adevelopers\/query-(\h+)-(\d+)-(\d+)\z/, developers.cache_key)
+      assert_match(/\Adevelopers\/query-(\h+)-(\d+)-(\d+)-(\d+)\z/, developers.cache_key)
 
-      /\Adevelopers\/query-(\h+)-(\d+)-(\d+)\z/ =~ developers.cache_key
+      /\Adevelopers\/query-(\h+)-(\d+)-(\d+)-(\d+)\z/ =~ developers.cache_key
 
       assert_equal ActiveSupport::Digest.hexdigest(developers.to_sql), $1
       assert_equal developers.count.to_s, $2
       assert_equal last_developer_timestamp.to_s(ActiveRecord::Base.cache_timestamp_format), $3
+      assert_equal first_developer_timestamp.to_s(ActiveRecord::Base.cache_timestamp_format), $4
     end
 
     test "cache_key for relation with includes" do
       comments = Comment.includes(:post).where("posts.type": "Post")
-      assert_match(/\Acomments\/query-(\h+)-(\d+)-(\d+)\z/, comments.cache_key)
+      assert_match(/\Acomments\/query-(\h+)-(\d+)-(\d+)-(\d+)\z/, comments.cache_key)
     end
 
     test "cache_key for loaded relation with includes" do
       comments = Comment.includes(:post).where("posts.type": "Post").load
-      assert_match(/\Acomments\/query-(\h+)-(\d+)-(\d+)\z/, comments.cache_key)
+      assert_match(/\Acomments\/query-(\h+)-(\d+)-(\d+)-(\d+)\z/, comments.cache_key)
     end
 
     test "it triggers at most one query" do
@@ -138,7 +148,7 @@ module ActiveRecord
 
     test "collection proxy provides a cache_key" do
       developers = projects(:active_record).developers
-      assert_match(/\Adevelopers\/query-(\h+)-(\d+)-(\d+)\z/, developers.cache_key)
+      assert_match(/\Adevelopers\/query-(\h+)-(\d+)-(\d+)-(\d+)\z/, developers.cache_key)
     end
 
     test "cache_key for loaded collection with zero size" do
@@ -157,19 +167,19 @@ module ActiveRecord
     test "cache_key with a relation having selected columns" do
       developers = Developer.select(:salary)
 
-      assert_match(/\Adevelopers\/query-(\h+)-(\d+)-(\d+)\z/, developers.cache_key)
+      assert_match(/\Adevelopers\/query-(\h+)-(\d+)-(\d+)-(\d+)\z/, developers.cache_key)
     end
 
     test "cache_key with a relation having distinct and order" do
       developers = Developer.distinct.order(:salary).limit(5)
 
-      assert_match(/\Adevelopers\/query-(\h+)-(\d+)-(\d+)\z/, developers.cache_key)
+      assert_match(/\Adevelopers\/query-(\h+)-(\d+)-(\d+)-(\d+)\z/, developers.cache_key)
     end
 
     test "cache_key with a relation having custom select and order" do
       developers = Developer.select("name AS dev_name").order("dev_name DESC").limit(5)
 
-      assert_match(/\Adevelopers\/query-(\h+)-(\d+)-(\d+)\z/, developers.cache_key)
+      assert_match(/\Adevelopers\/query-(\h+)-(\d+)-(\d+)-(\d+)\z/, developers.cache_key)
     end
 
     test "cache_key should be stable when using collection_cache_versioning" do
@@ -188,13 +198,15 @@ module ActiveRecord
       with_collection_cache_versioning do
         developers = Developer.where(salary: 100000).order(updated_at: :desc)
         last_developer_timestamp = developers.first.updated_at
+        first_developer_timestamp = developers.last.updated_at
 
-        assert_match(/(\d+)-(\d+)\z/, developers.cache_version)
+        assert_match(/(\d+)-(\d+)-(\d+)\z/, developers.cache_version)
 
-        /(\d+)-(\d+)\z/ =~ developers.cache_version
+        /(\d+)-(\d+)-(\d+)\z/ =~ developers.cache_version
 
         assert_equal developers.count.to_s, $1
         assert_equal last_developer_timestamp.to_s(ActiveRecord::Base.cache_timestamp_format), $2
+        assert_equal first_developer_timestamp.to_s(ActiveRecord::Base.cache_timestamp_format), $3
       end
     end