### Summary
This PR fixes an old problem with the caching of queries with limit.
The correction can be tested using the sample app submitted by @marckohlbrugge in #37555 (the most recent issue on the subject).

Before this change the cache_version format would be:

**"count-max_timestamp (usually updated_at)"**
ex: **"3-20191111224826321158"**

But in [some scenarios](https://github.com/rails/rails/issues/37555) involving limit and specific ordering
this data wasn't enough to invalidate the cache correctly, so
this PR changes the format to:

**"count-max_timestamp-min_timestamp"**
ex: **"3-20191111224826321158-20191111224826321030"**

### Benchmarks
The benchmarks were conducted using this script:
https://gist.github.com/victor-am/2a4302c8878b56221aceb613d001609c

Any suggestions on different scenarios for the benchmarks would be appreciated 😄 

**❗️Since this problem seems to occur almost exclusively with limit queries we could only apply this change on those scenarios to avoid messing with the non-limit queries. The downside would be to have two formats of cache_version (`count-timestamp` and `count-timestamp-timestamp`). Let me know if this is preferable 🤔**

And the results are as following:
```
========================= Query with limit 3 and order =========================

Warming up --------------------------------------
       Before change     1.000  i/100ms
        After change     1.000  i/100ms
Calculating -------------------------------------
       Before change     12.590  (±23.8%) i/s -     59.000  in   5.036258s
        After change     13.194  (± 7.6%) i/s -     66.000  in   5.037225s

Comparison:
        After change:       13.2 i/s
       Before change:       12.6 i/s - same-ish: difference falls within error


====================== Query with limit 100_000 and order ======================

Warming up --------------------------------------
       Before change     1.000  i/100ms
        After change     1.000  i/100ms
Calculating -------------------------------------
       Before change      6.656  (± 0.0%) i/s -     34.000  in   5.120976s
        After change      6.204  (±16.1%) i/s -     31.000  in   5.060640s

Comparison:
       Before change:        6.7 i/s
        After change:        6.2 i/s - same-ish: difference falls within error


============================== Query with limit 3 ==============================

Warming up --------------------------------------
       Before change   242.000  i/100ms
        After change   240.000  i/100ms
Calculating -------------------------------------
       Before change      2.582k (± 9.2%) i/s -     12.826k in   5.026062s
        After change      1.955k (±12.5%) i/s -      9.600k in   5.002996s

Comparison:
       Before change:     2582.0 i/s
        After change:     1954.6 i/s - 1.32x  slower


=========================== Query with limit 100_000 ===========================

Warming up --------------------------------------
       Before change     5.000  i/100ms
        After change     5.000  i/100ms
Calculating -------------------------------------
       Before change     62.924  (±14.3%) i/s -    310.000  in   5.018337s
        After change     54.579  (± 7.3%) i/s -    275.000  in   5.064948s

Comparison:
       Before change:       62.9 i/s
        After change:       54.6 i/s - same-ish: difference falls within error


======================= Query without limit (30 results) =======================

Warming up --------------------------------------
       Before change   204.000  i/100ms
        After change   154.000  i/100ms
Calculating -------------------------------------
       Before change      1.826k (±31.1%) i/s -      7.752k in   5.106217s
        After change      2.054k (±17.0%) i/s -      9.856k in   5.012670s

Comparison:
        After change:     2054.1 i/s
       Before change:     1826.0 i/s - same-ish: difference falls within error


==================== Query without limit (500_000 results) =====================

Warming up --------------------------------------
       Before change     1.000  i/100ms
        After change     1.000  i/100ms
Calculating -------------------------------------
       Before change     20.596  (± 9.7%) i/s -    103.000  in   5.036099s
        After change     19.041  (± 5.3%) i/s -     95.000  in   5.012132s

Comparison:
       Before change:       20.6 i/s
        After change:       19.0 i/s - same-ish: difference falls within error
```