@alipman88 2019-11-13
I like the straightforwardness of this approach, but I don't think it's quite robust enough.

Consider the following table, say for an Employee model:

```
SELECT * FROM employees;
+----+-------+---------------------+---------------------+
| id | name  | created_at          | updated_at          |
+----+-------+---------------------+---------------------+
|  1 | alice | 2019-11-01 00:00:00 | 2019-11-01 00:00:00 |
|  2 | carol | 2019-11-02 00:00:00 | 2019-11-02 00:00:00 |
|  3 | doug  | 2019-11-03 00:00:00 | 2019-11-03 00:00:00 |
|  4 | bob   | 2019-11-04 00:00:00 | 2019-11-04 00:00:00 |
+----+-------+---------------------+---------------------+
```

Suppose we selected the first three employees, sorted alphabetically:

```
Employee.order(:name).limit(3)
# alice, bob & carol
```

The minimum and maximum `updated_at` timestamps of the above records would be `2019-11-01 00:00:00` (alice) and `2019-11-04 00:00:00` (bob).

Now, imagine we delete carol from the database:

```
Employee.find_by(name: "carol").destroy
```

If we run our original query, we get different employees:

```
Employee.order(:name).limit(3)
# alice, bob & doug
```

However, the minimum and maximum `updated_at` timestamps are unchanged, thus the `cache_version` will stay the same.