@victor-am 2019-11-13
Yeah, there is still this scenario left :/

I went through this route to avoid changing much of the current cache_version implementation, but I agree it would be better if we could solve all cases altogether 🤔 

I see you suggested 2 approaches in the issue (https://github.com/rails/rails/issues/37555#issuecomment-552237435) gonna experiment with those and benchmark them to see if they can better solve the problem 😄