@victor-am 2019-11-13
# AR cache invalidation
#open-source/rails

So, to make the tests easier I've come up with this script to test the cache invalidation in each approach:
https://gist.github.com/victor-am/6558e38342ea990d4a6c7d2e7f87671a

Basically I've described those 3 scenarios:

**Scenario A:**
Replacing the record with highest timestamp from the results 
```
  # QUERY RESULTS FOR SCENARIO
  # (timestamps relative to @epoch)
  #
  # | Before              | After              |
  # |--------|------------|-------|------------|
  # | id     | updated_at | id    | updated_at |
  # |--------|------------|-------|------------|
  # | 3      | + 3 days   | 4     | + 4 days   |
  # | 2      | + 2 days   | 3     | + 3 days   |
  # | 1      | + 1 days   | 2     | + 2 days   |
```

**Scenario B:**
Replacing the record with the middle timestamp from the results
```
  # QUERY RESULTS FOR SCENARIO
  # (timestamps relative to @epoch)
  #
  # | Before              | After              |
  # |--------|------------|-------|------------|
  # | id     | updated_at | id    | updated_at |
  # |--------|------------|-------|------------|
  # | 3      | + 3 days   | 3     | + 3 days   |
  # | 2      | + 2 days   | 4     | + 2 days   |
  # | 1      | + 1 days   | 1     | + 1 days   |
```

**Scenario C:**
Replacing the record with the lowest timestamp from the results
```
  #
  # QUERY RESULTS FOR SCENARIO
  # (timestamps relative to @epoch)
  #
  # | Before              | After              |
  # |--------|------------|-------|------------|
  # | id     | updated_at | id    | updated_at |
  # |--------|------------|-------|------------|
  # | 3      | + 3 days   | 3     | + 3 days   |
  # | 2      | + 2 days   | 2     | + 2 days   |
  # | 1      | + 1 days   | 4     | + 0 days   |
```

## So about the solutions:

### Last ID
I think going with the last timestamp would be better here since relying on an id field adds one more requirement for the cache version generation, while getting the last timestamp should achieve almost the same results.

The problem is that it could still fail in Scenario B (same as my approach in #37694 does) depending on the order criteria used.

### ID Checksum
The problem I see with this approach is that it adds quite some complexity (SQL-wise, in creating a dependency with the id field and in the fact that it would have to be implemented in an opt-in basis because of the performance hit).

Although it seems to be the only one to solve all three scenarios that we came up with for now 🤔

It's important to notice that this doesn't solve all potential cache invalidation problems.

For example, if the collection items didn't change but their order did (without touching the updated_at of any of the items). Although it's a far more unlikely scenario than Scenario B for sure.

### Adding the MIN(timestamp) to cache_version
This was my proposed solution at #37694 and more or less fall into the same pros/cons of the **Last ID** solution.

## Conclusion
Looking at these options I can see the pros and cons of each approach but I'm not sure which trade-off the project rather buy (solve more cases with costs in performance/complexity or solve less cases with much lower perf/complexity costs).

I guess a rails team member could help shed some light on this issue 💡 

Btw @alipman88 , did my rationale made sense? Did I miss something? 🤔